Move upnp methods to own file
